refactor: split seed_data_service into a package
894-line monolith → seed_data/ package with focused modules:
constants, event_generators, health_score_generators, support_generators, service.
seed_data_service.py kept as a thin re-export for backward compat.
