+++ backend/app/services/seed_data/__init__.py
@@ -1,0 +1,3 @@
+from .service import seed_data_service
+
+__all__ = ["seed_data_service"]
